separate files
- extract get_single_project
- extract remove_project_tags
- update import path
- update readme
